fix: rename channels to notifications
